backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
python-jose[cryptography]